requirements.txt
attrs==19.3.0
Babel==2.7.0
Django==2.2.7
django-phonenumber-field==3.0.1
django-tables2==2.2.1
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
phonenumbers==8.10.23
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
pytest-django==3.7.0
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.6.0
